==> code_redacted/00_base_cleaning_and_geocoding.ipynb <==
import pandas as pd
import warnings
from geocode import concat_full_address, fetch_geodata

Load the data from `battery.csv` into a pandas data frame. 

# importing raw data. This is the original excel data grs provided us exported as csv.
raw_df = pd.read_csv('../data/battery.csv', sep=';')

df = raw_df.copy()
df.head() # cell outputs cleared

df.shape # shape redacted

# Check data info and show the data type of each column
print(df.info()) # description redacted

## Fix column names formats

We want to change the empty spaces to underscores. Also we want to use only low letters in the column names.

# replacing empty spaces with underscores in column names
df.columns = df.columns.str.replace(' ','_')
df.columns # description redacted

# changing all column names to lower case
df.columns = df.columns.str.lower()
df.columns # description redacted

## Imputing Latitude and Longitude
Test run with few data :)

warnings.simplefilter('ignore')
df_test = df.iloc[:10]
concat_full_address(df_test, address_col='volle_Addresse')
fetch_geodata(df_test, address_col='volle_Addresse')

df_test.head(3) # output redacted

# check if address data has missing values
df_test[['stasse', 'plz', 'ort', 'bundesland']].isna().sum()

Fetching  lat/long geocodes for all addresses. This will run 2-3 hours.

FETCH_ALL = False

if FETCH_ALL:
    concat_full_address(df, address_col='volle_Addresse')
    fetch_geodata(df, address_col='volle_Addresse')

df[['stasse', 'plz', 'ort', 'bundesland']].isna().sum()

# pickle will preserve the full state of a pandas dataframe including dtypes
if FETCH_ALL:
    df.to_pickle("../data/battery_with_geo.pkl")
