==> code_redacted/01_data_cleaning.ipynb <==
## Preliminary data cleaning

import numpy as np
import pandas as pd
import missingno as msno
import seaborn as sns
import matplotlib.pyplot as plt
import datetime

Load the data from `battery.csv` into a pandas data frame. 

df = pd.read_pickle("../data/battery_with_geo.pkl")

df.head() # redacted

df.shape # redacted

df.columns # redacted

# Check data info and show the data type of each column
print(df.info()) # redacted

## Dealing with missing values

df.isna().sum() # redacted

df.nunique() # redacted

msno.matrix(df) # redacted

we see a pattern:
- **IF** disponiert_am is missing **THEN** also abholdatum is missing
- both columns have *almost* the same missing values
  - *in some few cases* abholdatum has a missing value where disponiert_am has a value

# nice one liner for statistical overview over missing values
df.isna().mean().mul(100).sort_values(ascending=False).iloc[:15]

### Imputing constant values for the missing values
We do this only for non-date columns. date columns are handled in "Data Types and Transforming Data"

#impute "unknown" for  string columns
df[["name_2", "stasse", "name_1",]] = df[["name_2", "stasse", "name_1",]].fillna("unknown")

#impute "other" for categorical columns
df[["typ", "gelieferter_behältertyp"]] = df[["typ", "gelieferter_behältertyp"]].fillna("other")

#impute 0 for numeric columns
df[["kreisgemeindeschlüssel", "länderschlüssel", "region"]] = df[["kreisgemeindeschlüssel", "länderschlüssel", "region"]].fillna(0)

## Dealing with dublicates

# check how many duplicated rows exist in the data frame
df.duplicated().value_counts() # redacted

# remove duplicates
df = df.drop_duplicates()
# reset index inplace
df.reset_index(inplace=True, drop=True)
df.head(5) # redacted

# check again for missing values
df.isna().sum() # redacted

#print the total number of missing values
print(df.isnull().values.sum()) # redacted

print(df.name_2.value_counts())
print("count of missing values in name_2: ", df.name_2.isna().sum()) # redacted

# Übersicht über gelieferte Behältertypen
print(df.gelieferter_behältertyp.value_counts())
print("count of missing values in gelieferter_behältertyp: ", df.gelieferter_behältertyp.isna().sum()) # redacted

df['gelieferter_behältertyp'].isnull() # redacted

As we can see Pandas recognized the "0s" as a missing value. 

## Data Types and Transforming Data

In the following we want to get the data types into the right shape. The dtype `object` for ecample means that there are several data types in the corresponding column.

# check data types in data frame
df.dtypes # redacted

# Select numeric columns
df.select_dtypes('number') # redacted

df = df.astype({'kreisgemeindeschlüssel': int})

df['nettogewicht_in_kg'] = df['nettogewicht_in_kg'].apply(lambda x: x.replace(',','.'))

df = df.astype({'nettogewicht_in_kg': float})
df['nettogewicht_in_kg'] = df['nettogewicht_in_kg'].apply(lambda x: round(x))
df = df.astype({'nettogewicht_in_kg': int})

# type of first date entry
type(df['auftragsdatum'][0])

# type of first date entry
type(df['auftrag_bestätigt_am'][0])

As you can see our date entry is just a string. We can change that to a date time very easy as we already learned. Note that you also could do that in the beginning, when you read in the csv file with the parameter `parse_date=['date']`.

# change "date" dtype to datetime with format %Y/%m/%d
df['auftragsdatum'] = pd.to_datetime(df['auftragsdatum'], format='%d.%m.%y')

# type of first date entry
type(df['auftragsdatum'][0])

df_datum = pd.DataFrame(df[['auftragsdatum', 'auftrag_bestätigt_am', 'disponiert_am', 'abholdatum', 'zurückgemeldet_am']])
df_datum.head()

df['auftrag_bestätigt_am'] = df['auftrag_bestätigt_am'].replace('-', np.nan)
df['auftrag_bestätigt_am'] = df['auftrag_bestätigt_am'].fillna("01.01.99")

df['auftrag_bestätigt_am'] = pd.to_datetime(df['auftrag_bestätigt_am'], format='%d.%m.%y')

df['disponiert_am'] = df['disponiert_am'].fillna("01.01.99")

df['disponiert_am'] = pd.to_datetime(df['disponiert_am'], format='%d.%m.%y')

df['abholdatum'] = df['abholdatum'].fillna("01.01.99")

df['abholdatum'] = pd.to_datetime(df['abholdatum'], format='%d.%m.%y')

df['zurückgemeldet_am'] = df['zurückgemeldet_am'].fillna("01.01.99")

df['zurückgemeldet_am'] = pd.to_datetime(df['zurückgemeldet_am'], format='%d.%m.%y')

df_datum = pd.DataFrame(df[['auftragsdatum', 'auftrag_bestätigt_am', 'disponiert_am', 'abholdatum', 'zurückgemeldet_am']])
df_datum.head()

df.groupby(['auftrag_bestätigt_am']).size() # redacted

print(df['herkömliche_übergabestelle'].isnull().values.sum())

type(df['herkömliche_übergabestelle'][0])

# replace string by boolean
df['herkömliche_übergabestelle'] = df['herkömliche_übergabestelle'].map({'x':True}) 
type(df['herkömliche_übergabestelle'][0])

df['herkömliche_übergabestelle'].fillna(True, inplace=True)

print(df['herkömliche_übergabestelle'].isnull().values.sum())

# replace string by boolean
df['qualifizierte_annahmestelle'] = df['qualifizierte_annahmestelle'].map({'x':True, '-': False}) 
df['qualifizierte_annahmestelle'] # redacted

# replace string by boolean
df['qualifizierte_sammelstelle'] = df['qualifizierte_sammelstelle'].map({'x':True, '-': False}) 
df['qualifizierte_sammelstelle'] # redacted

# display number of distinct elements
print("The count of distinct categories in angeforderter_behältertyp is: ", df.angeforderter_behältertyp.nunique())
print("The count of distinct categories in gelieferter_behältertyp is: ", df.gelieferter_behältertyp.nunique())

# Descriptive statistics for column age_group
print(df.angeforderter_behältertyp.value_counts())
# redacted

print(df.gelieferter_behältertyp.value_counts())
# redacted

df.region.value_counts() # redacted

sorted_regions = df.region.unique()
sorted_regions.sort()
sorted_regions

# drop columns
df.drop('los_1p7', axis='columns', inplace=True)
df.drop('auftragsjahr', axis='columns', inplace=True)
df.drop('auftragsmonat', axis='columns', inplace=True)
df.drop('länderschlüssel', axis='columns', inplace=True)
df.drop('region', axis='columns', inplace=True)
df.drop('disponiert_am', axis='columns', inplace=True)
df.drop('kategorie', axis='columns', inplace=True)

# check data types in data frame
df.dtypes # redacted

## Dealing with Outliers
In the following we want to detect and handle outliers. 

An outlier is an exceptionally high or low value. Based on this definition, a first idea to detect outliers would be to simply cut down the highest and lowest points of the dataset.

To create a better understanding, lets look only the numerical values. 

# select numeric columns
df_numeric = pd.DataFrame(df[['bruttogewicht_in_kg', 'nettogewicht_in_kg', 'angemeldete_containeranzahl', 'angeforderter_behältertyp']])


df_numeric.head()

df_numeric.describe().T # redacted

What we can observe from reading this data:
- `bruttogewicht_in_kg`: very high (# redacted) 'bruttogewicht' 
- `nettogewicht_in_kg`: we can see few outliers that 1) have negative weight and 2) very high 'nettogewicht' (# redacted)
- `angemeldete_containeranzahl`: very high 'contraineranzahl' (# redacted) (outlier)

Dropping the outliers.

# drop rows with "brutto_gewicht_in_kg" larger than 30000
df = df.drop(df[df.bruttogewicht_in_kg > 30000].index)

# drop rows with "nettogewicht_in_kg" larger than 25000
df = df.drop(df[df.nettogewicht_in_kg > 25000].index)

# drop all rows with negative "nettogewicht_in_kg" 
df = df.drop(df[df.nettogewicht_in_kg < 0.0].index)

# drop the rest non-relevant values
#df = df.drop(df[(df.bruttogewicht_in_kg > 15000) & (df.angemeldete_containeranzahl == 10)].index)

# drop rows with "angemeldete_containeranzahl" larger than 300
df = df.drop(df[df.angemeldete_containeranzahl > 300].index)

# drop rows with "angemeldete_containeranzahl" = 0
df = df.drop(df[df.angemeldete_containeranzahl == 0].index)

# drop rows with "auftragsstatus" = 'Storniert'
df = df.drop(df[df.auftragsstatus == 'Storniert'].index)

# drop rows with "auftragsstatus" = 'Fehlfahrt'
df = df.drop(df[df.auftragsstatus == 'Fehlfahrt'].index)

df = df.drop(df[df.nettogewicht_in_kg == 1].index)

df = df.drop(df[df.bruttogewicht_in_kg == 3].index)

df.query("bruttogewicht_in_kg == nettogewicht_in_kg") # redacted

df.bundesland.value_counts() # redacted

df.bundesland = np.where(((df.plz == 27639) & (df.bundesland == '?')), 'Niedersachsen', df.bundesland)

df.bundesland = np.where(((df.plz == 57234) & (df.bundesland == '?')), 'Nordrhein-Westfalen', df.bundesland)

df.bundesland = np.where(((df.plz == 33333) & (df.bundesland == '?')), 'Nordrhein-Westfalen', df.bundesland)

print("Shape raw data: ", raw_df.shape)
print("Shape cleaned outliers in the numerical values: ", df.shape)
print("Shape cleaned data without missing values and outliers in weights: ", df.shape)
# redacted

# show min and max netto and brutto of the cleaned data, respectively
print("min netto: ", df.nettogewicht_in_kg.min())
print("max netto: ", df.nettogewicht_in_kg.max())
print()
print("min brutto: ", df.bruttogewicht_in_kg.min())
print("max brutto: ", df.bruttogewicht_in_kg.max())
# redacted

df.info()
# redacted

df.isna().sum() # redacted

df[df.volle_Addresse.isna()] # redacted

2 addresses are not complete so there's also not lat, long.

# pickle will preserve the full state of a pandas dataframe including dtypes
df.to_pickle("../data/battery_cleaned_with_geo.pkl")
