==> code_redacted/04_testing.ipynb <==
import numpy as np
import pandas as pd
from datetime import datetime as dt
import time
import math
import datetime

import prediction

### Testing filtering of single pick-ups and single initial deliveries

create test data

orders = pd.read_pickle("../data/battery_cleaned_with_geo.pkl")
orders = orders[["übergabestellennummer", "auftragsnummer", "auftragsstatus", "auftragstyp", "abholdatum", "nettogewicht_in_kg", "angemeldete_containeranzahl", "angeforderter_behältertyp", "gelieferter_behältertyp", "gelieferte_behälteranzahl"]]

# create test data sets

#redacted

test = pd.read_pickle("../data/test_103509220.pkl")
test.drop(test.query("auftragsnummer == 23615989").index, inplace=True)
test.to_pickle("../data/test_erstausstattung_abholauftrag_both_single.pkl")

# has  one order of type initial delivery and one pick-up order. This should be filtered out.
test_orders_in = pd.read_pickle("../data/test_erstausstattung_abholauftrag_both_single.pkl")
test_orders_out = pd.read_pickle("../data/test_erstausstattung_abholauftrag_both_single.pkl")
assert prediction.drop_single_pick_ups_and_single_initial_deliveries(test_orders_in).shape[0] == 0

# 1713671 has a long list of orders including an initial delivery somewhere in between.  This is fine - nothing should change here.
test_orders_in = pd.read_pickle("../data/test_1713671.pkl")
test_orders_out = pd.read_pickle("../data/test_1713671.pkl")
assert prediction.drop_single_pick_ups_and_single_initial_deliveries(test_orders_in).equals(test_orders_out)

# 103509220 has order type initial delivery as first order and 2 later pick-up orders. This is fine - nothing should change here.
test_orders_in = pd.read_pickle("../data/test_103509220.pkl")
test_orders_out = pd.read_pickle("../data/test_103509220.pkl")
assert prediction.drop_single_pick_ups_and_single_initial_deliveries(test_orders_in).equals(test_orders_out)

# 100222 has only one order of type initial delivery. This should be filtered out.
test_orders_in = pd.read_pickle("../data/test_100222.pkl")
test_orders_out = pd.read_pickle("../data/test_100222.pkl")
assert prediction.drop_single_pick_ups_and_single_initial_deliveries(test_orders_in).shape[0] == 0

# collection point 107573 has 2 orders with same date but different container type. we can not calculate a prediction.
test_orders_in = pd.read_pickle("../data/test_107573.pkl")
test_orders_out = pd.read_pickle("../data/test_107573.pkl")
assert prediction.drop_single_pick_ups_and_single_initial_deliveries(test_orders_in).shape[0] == 0

# 100222
test_orders_in = pd.read_pickle("../data/test_100222.pkl")
test_orders_out = pd.read_pickle("../data/test_100222.pkl")
assert prediction.filter_for_min3_pick_ups_and_initial_deliveries(test_orders_in).shape[0] == 0

test_orders_in = pd.read_pickle("../data/test_103509220.pkl")
test_orders_out = pd.read_pickle("../data/test_103509220.pkl")
assert prediction.filter_for_min3_pick_ups_and_initial_deliveries(test_orders_in).shape[0] == 0

test_orders_in = pd.read_pickle("../data/test_1713671.pkl")
test_orders_out = pd.read_pickle("../data/test_1713671.pkl")
assert prediction.filter_for_min3_pick_ups_and_initial_deliveries(test_orders_in).equals(test_orders_out)

temp = pd.read_pickle("../data/test_1713671.pkl")
temp = temp.iloc[:4]
temp.to_pickle("../data/test_1713671_delivery_latest.pkl")

test_orders_in = pd.read_pickle("../data/test_1713671_delivery_latest.pkl")
test_orders_out = pd.read_pickle("../data/test_1713671_delivery_latest.pkl")
assert prediction.filter_for_min3_pick_ups_and_initial_deliveries(test_orders_in).shape[0] == 3


# 111930
test_orders_in = pd.read_pickle("../data/test_111930.pkl")
test_orders_out = pd.read_pickle("../data/test_111930.pkl")
assert prediction.filter_for_min3_pick_ups_and_initial_deliveries(test_orders_in).shape[0] > 2
assert prediction.filter_for_min3_pick_ups_and_initial_deliveries(test_orders_in).iloc[-1].auftragstyp == "Abholauftrag"

# collection point 107573 has 2 orders with same date but different container type. we can not calculate a prediction.
test_orders_in = pd.read_pickle("../data/test_107573.pkl")
test_orders_out = pd.read_pickle("../data/test_107573.pkl")
assert prediction.filter_for_min3_pick_ups_and_initial_deliveries(test_orders_in).shape[0] == 0
